Run validation as a unit test.
Skip it on non-ICU environments
